+++ packages/terra-form-input/CHANGELOG.md
@@ -10,6 +10,11 @@ Unreleased
 * Placeholder text styling
 * Webdriver and dev site tests for disabled, placeholder, and value styling
 
+1.22.0 - (August 29, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 1.21.0 - (August 8, 2018)
 ------------------
 ### Changed